examples/fastapi-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
auth-kit-fastapi==1.0.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
redis==5.0.1